extract toolkit select component
